Rename: config 파일명 수정
